% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci-calculation.R
\name{ci}
\alias{ci}
\title{Confidence interval for a proportion}
\usage{
ci(num, denom)
}
\arguments{
\item{num}{This is the numerator of the proportion.}

\item{denom}{This is the denominator of the proportion.}
}
\value{
The lower and upper bounds of the 95\% confidence interval for a given proportion.
}
\description{
Use this function to calculate the lower and upper 95\% confidence intervals for a proportion (e.g. rate) using the normal approximation method. The confidence interval estimates a range of values that capture the true population value for a variable of interest. For a 95\% confidence interval, if a rate of 5.5 falls within a 95\% confidence interval of 4.5, 6.5, it means that one is 95\% confidence that the true value from the population falls within that range.
}
\examples{

ci(5, 25)


}
