% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life-expectancy.R
\name{life_table}
\alias{life_table}
\title{Life table calculations to be used in \code{make_life_table()}. This function isn't exported in the package.}
\usage{
life_table(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a data frame
}
\description{
this is the function called in the \code{make_life_table()}. It calculates the vectors required for creating life expectancy methods outlined in the Public Health England calculator that lives \href{https://fingertips.phe.org.uk/documents/phe\%20life\%20expectancy\%20calculator.xlsm}{here}.
}
\details{
Some things have been added, including the option for 0-4 ages (instead of 0 and 1-4) and 85+ (instead of 85-89 and 90+)
}
\examples{
# don't run this on its own - run make_life_table()
}
