% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-smc-fonts.R
\name{load_smc_fonts}
\alias{load_smc_fonts}
\title{Download fonts needed to format charts with SMC style guide}
\usage{
load_smc_fonts()
}
\value{
this will enable all of the \code{extrafont} fonts on your computer.
}
\description{
This function uses the 'extrafont' package to download and enable Arial and Georgia which are the official fonts for the SMC style guide.
}
\examples{
# just run the function empty
}
