% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rse-calculation.R
\name{prop_rse}
\alias{prop_rse}
\title{Relative standard error calculation for sample and survey data}
\usage{
prop_rse(prop_sd, n, method)
}
\arguments{
\item{prop_sd}{This is the standard deviation of the proportion.}

\item{n}{This is the sample size of the data.}

\item{method}{This is the method by which the RSE is calculated.
\itemize{
\item \code{"sample"}: used to calculate the RSE of surveillance data sets
\item \code{"survey"}: utilized to calculate the RSE of data from a survey or a limited data set from a larger population.
}}
}
\value{
The relative standard error of sample or survey data.
}
\description{
Use this function to calculate the relative standard error (RSE) for a sample and survey data. The RSE measures the reliability of an estimate by dividing the standard error of the estimate by the estimate. The RSE is displayed as a percentage of the estimate.
}
\examples{
\dontrun{

prop_rse(25, 1050, method = "survey")
}
}
