% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-calculation.R
\name{prop_sd}
\alias{prop_sd}
\title{Standard deviation calculation for a proportion}
\usage{
prop_sd(num, denom)
}
\arguments{
\item{num}{This is the numerator of the proportion.}

\item{denom}{This is the denominator of the proportion.}
}
\value{
The standard deviation of a proportion.
}
\description{
Use this function to calculate the standard deviation of a proportion (e.g. rate). The standard deviation is a measure of spread that represents how far the population proportion is from the sample proportions on average.
}
\examples{
prop_sd(5,20)

}
