% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-calredie-tsv.R
\name{read_tsv_cr}
\alias{read_tsv_cr}
\title{Using `read_tsv_cr`}
\usage{
read_tsv_cr(file)
}
\arguments{
\item{file}{this should be a string}
}
\value{
a data frame
}
\description{
This is a wrapper for `read.csv2` that is meant specifically for CalREDIE tab separated files from the DDP.
}
\examples{
\dontrun{
data <- read_tsv_cr("file.tsv")
}
}
