% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc-geographies.R
\name{smc_city_clean}
\alias{smc_city_clean}
\title{Cleaning up San Mateo Cities in a data frame}
\usage{
smc_city_clean(data,
      city_col = "city",
      new_col = "city_clean")
}
\arguments{
\item{data}{This is the name of the dataframe with the city variable in it that you'd like to clean}

\item{city_col}{(optional): This is a string that specifies the name of the city variable you want to clean. By default, the function assumes the variable is called \code{city}.}

\item{new_col}{(optional): This is a string to specify the name of the variable with the cleaned city names. By default the cleaned cities will be saved in a variable called \code{city_clean}}
}
\value{
a dataset with a new variable for the cleaned city values
}
\description{
This function is meant to be used to clean up San Mateo cities in a dataframe. It will only look for San Mateo County cities - it doesn't look for other cities. Non-San Mateo cities will return an \code{NA} in the \code{city_clean} column.
}
\examples{
data <- data.frame(
 city_dirty = c("Burligame", "Fost City", "San Mato", "Daily Cit", "S S Francisco",
                "South San Fransico", "SoSan Franc", "San Francisco")
 )

data \%>\%
    smc_city_clean(city_col = "city_dirty",
                   new_col = "city_smc")


}
