% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc-geographies.R
\name{smc_zip_region_sort}
\alias{smc_zip_region_sort}
\title{Classify zip codes by region in San Mateo County}
\usage{
city_clean(data,
    city_col,
    new_col)
}
\arguments{
\item{data}{This is the name of the dataframe with the zip code variable in it that you'd like to sort}

\item{zip_col}{(optional): This is a string that specifies the name of the zip code variable you want to sort. By default the function assumes the zipcodes are stored in a variable called \code{zip}}

\item{region_col}{(optional): This is a string to specify the name of the variable with the zip regions. By default the sorted zipcodes will be saved in a variable called \code{region_col}}
}
\value{
a dataset with a new variable for the sorted zip codes
}
\description{
This function sorts 5-digit San Mateo zip codes into county regions: north, south, mid, coastside and "not a residential zip". Non-San Mateo zipcodes will return an \code{NA} in the \code{city_clean} column.
}
\examples{
\dontrun{
 data1 <- data \%>\%
    smc_zip_region_sort(zip_col = "zip",
                        region_col = "zip_region")

}


}
