% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcepi-package.R
\docType{package}
\name{smcepi-package}
\alias{smcepi}
\alias{smcepi-package}
\title{smcepi: Useful functions for San Mateo County Epidemiologists}
\description{
Handy functions for use by San Mateo County epidemiologists.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/San-Mateo-County-Health-Epidemiology/smcepi}
  \item \url{https://San-Mateo-County-Health-Epidemiology.github.io/smcepi/}
  \item \url{https://san-mateo-county-health-epidemiology.github.io/smcepi/}
  \item Report bugs at \url{https://github.com/San-Mateo-County-Health-Epidemiology/smcepi/issues}
}

}
\keyword{internal}
