% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-smc.R
\name{theme_ft_smc}
\alias{theme_ft_smc}
\title{Formatting \code{flextable} tables according to our SMC style guide}
\usage{
theme_ft_smc(ft)
}
\arguments{
\item{ft}{This should be a \code{flextable} object}
}
\value{
a \code{flextable} object with custom formatting
}
\description{
This function will format a flextable object according to the San Mateo County Office of Epidemiology and Evaluation's style guide.
}
\examples{
\dontrun{

library(flextable)
library(dplyr)

iris \%>\%
   slice(1:5) \%>\%
   select(Species, everything()) \%>\%
   flextable() \%>\%
   footnote(i = 1, j = 1, value = as_paragraph("footer test")) \%>\%
   theme_ft_smc()
}

}
