% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-smc.R
\name{theme_gg_smc}
\alias{theme_gg_smc}
\title{Formatting \code{ggplot2} charts according to our SMC style guide}
\usage{
theme_gg_smc(plot,
  plot_lines = "horizontal",
  legend_loc = "top",
  title_font_size = 18,
  subtitle_font_size = 12,
  caption_font_size = 8,
  axis_font_size = 10)
}
\arguments{
\item{plot}{This should be a \code{ggplot2} object}

\item{plot_lines}{Specify which lines should appear on your chart. The default is horizontal lines, but you can also choose:
\itemize{
\item \code{"horizontal"} (the default): only horizontal lines
\item \code{"vertical"}: only vertical lines
\item \code{"both"}: horizontal and vertical lines
\item \code{"none"}: no lines on the plot
}}

\item{legend_loc}{Specify the legend location. The default is for the legend to appear at the top, but you can override this. The available options are: “left”,“top”, “right”, “bottom” and "none"}

\item{title_font_size}{A numeric value to specify the font size of the title. The default is 18, the minimum is 6 and the maximum is 60.}

\item{subtitle_font_size}{A numeric value to specify the font size of the subtitle. The default is 12, the minimum is 6 and the maximum is 60.}

\item{caption_font_size}{A numeric value to specify the font size of the caption. The default is 8, the minimum is 6 and the maximum is 60.}

\item{axis_font_size}{A numeric value to specify the font size of the axis labels. The default is 10, the minimum is 6 and the maximum is 60.}
}
\value{
a ggplot2 object with custom formatting
}
\description{
This function is meant to be used with \code{ggplot2} charts to quickly format them according to the San Mateo County Office of Epidemiology and Evaluation's style guide.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
# creating a jitter plot with `theme_gg_smc`
\dontrun{
library(ggplot2)

iris \%>\%
  ggplot(aes(x = Species, y = Sepal.Length)) +
  geom_point(position = position_jitter(w = 1/6, h = 0),
             size = 4,
             color = "#006cb6",
             alpha = 2/3) +
  scale_y_continuous(limits = c(4, 8),
                     breaks = seq(4, 8, 2)) +
  labs(title = "Iris Jitter Plot") +
  theme_gg_smc()
}
}
