% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-smc.R
\name{theme_pl_smc}
\alias{theme_pl_smc}
\title{Formatting \code{plotly} charts according to our SMC style guide}
\usage{
theme_pl_smc(plot,
   plot_lines = "horizontal",
   legend_loc = "top",
   ystart = "tozero",
   title = NULL,
   x_lab = NULL,
   y_lab = NULL)
}
\arguments{
\item{plot}{A \code{plotly} plot object}

\item{plot_lines}{Specify which lines should appear on your chart. The default is horizontal lines, but you can also choose:
\itemize{
\item \code{"horizontal"} (the default): only horizontal lines
\item \code{"vertical"}: only vertical lines
\item \code{"both"}: horizontal and vertical lines
\item \code{"none"}: no lines on the plot
}}

\item{legend_loc}{Specify the legend location. The default is for the legend to appear at the top, but you can override this. The available options are: “left”,“top”, “right”, “bottom” and "none"}

\item{ystart}{Specify whether or not the y-axis should start at zero. The default is to start at zero. The options are:
\itemize{
\item \code{"tozero"} (the default) : axis starts at 0
\item \code{"none"}: axis starts wherever plotly decides
}}

\item{title}{a string of text for the plot title}

\item{x_lab}{a string of text for the x-axis label}

\item{y_lab}{a string of tect for the y-axis label}
}
\value{
a plotly object with custom formatting
}
\description{
This function is meant to be used with \code{plotly} charts to quickly format them according to the San Mateo County Office of Epidemiology and Evaluation's style guide
}
\examples{
\dontrun{
library(plotly)
library(dplyr)

title <- "Sepal Lengths of Iris Species"
x <- "Species"
y <- "Sepal Length"

iris_vals <- iris \%>\%
  distinct(Species) \%>\%
  mutate(number = as.numeric(Species))

iris \%>\%
  plot_ly() \%>\%
  add_markers(x = ~jitter(as.numeric(Species)), y = ~Sepal.Length,
              color = ~Species,
              marker = list(size = 6),
              hoverinfo = "text",
              text = ~paste0("Group: ", Species,
                             "<br>xval: ", Sepal.Length),
              showlegend = FALSE) \%>\%
              layout(xaxis = list(tickvals = iris_vals$number,
                                  ticktext = iris_vals$Species)) \%>\%
  theme_pl_smc(plot_lines = "vertical",
               title = title,
               x_lab = x,
               y_lab = y)
}
}
