% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-smc.R
\name{theme_smc}
\alias{theme_smc}
\title{Using `theme_smc`}
\usage{
theme_smc(plot, plot_lines = "horizontal", legend_loc = "top")
}
\arguments{
\item{plot}{this should be a ggplot2 object}

\item{plot_lines}{specify if you'd like horizontal, vertical, both (horizontal and vertical) or no (none) lines on the plot. the default is horizontal}

\item{legend_loc}{specify if you'd like the legend to appear at the top, bottom or to not appear at all (none)}
}
\description{
Using `theme_smc`
}
