% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azure-helpers.R
\name{varchar_max}
\alias{varchar_max}
\title{Get character variable (varchar) lengths for import into Azure}
\usage{
varchar_max(data)
}
\arguments{
\item{data}{a 2x2 dataset that you plan to import into Azure from R}
}
\value{
a named list that specifies the varchar(n) length for each variable
}
\description{
Use this to generate a list of varchar(`n`) specifications to be used in an R to Azure import. Each character variable is given a field type of varchar(`n`) where `n` is based on the maximum string length per variable. `n` is 255 for string lengths of 255 or less and is the maximum string length for the variable for any string longer than 255 (ex: `varchar(1001)`)
}
\examples{
data <- data.frame(
          col1 = c("blue", "pink", "green", "yellow"),
          col2 = c(paste(rep("a", 1000), collapse = ""), "b", "cc", "ddd"))

varchar_max(data)

# output:
#           col1           col2
# "varchar(255)" "varchar1000)"
}
