% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change-selected-text-font-color.R
\name{gg_color_title}
\alias{gg_color_title}
\title{Title}
\usage{
gg_color_title(title_text, highlight_words, highlight_colors)
}
\arguments{
\item{highlight_colors}{}
}
\description{
Title
}
