% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life-expectancy.R
\name{make_life_table}
\alias{make_life_table}
\title{Creating life tables}
\usage{
make_life_table(data,
                grouping_vars = NULL,
                age_cat_var = "age_cat",
                deaths_var = "deaths",
                population_var = "population")
}
\arguments{
\item{data}{a 2x2 data frame with variables for age categories, population years and death count.}

\item{grouping_vars}{a list of variables used to group the output.}

\item{age_cat_var}{the name of the variable with age categories if the variable has a name other than "age_cat"}

\item{deaths_var}{the name of the variable with death counts if the variable has a name other than "deaths"}

\item{population_var}{the name of the variable with the population years if the variable has a name other than "population"}
}
\value{
a data frame with a variable for each of the columns in the PHE life table
}
\description{
This function creates a life table based on the life table from Public Health England that lives \href{https://fingertips.phe.org.uk/documents/phe\%20life\%20expectancy\%20calculator.xlsm}{here}. It is designed to work with grouped data so that you can calculate life tables for multiple groups at once.
}
\examples{
# step 1: generate data from which to calculate life expectancy
test_data <- data.frame(
  year = c(rep(2020, 20), rep(2021, 20)),
  ages = rep(c("0", "1-4", "5-9", "10-14", "15-19","20-24", "25-29", "30-34", "35-39", "40-44",
  "45-49", "50-54", "55-59", "60-64", "65-69", "70-74", "75-79", "80-84", "85-89", "90+"), 2),
  deaths = c(206, 37, 23, 23, 105, 162, 268, 314, 413, 584, 954,
            1359, 1912, 2824, 4507, 5851, 7117, 8192, 7745, 6442,
            232, 30, 41, 22, 194, 168, 315, 313, 406, 643, 963,
            1446, 1979, 2814, 4587, 5874, 7111, 8221, 7825, 6540),
  population_est = c(50698, 215400, 280458, 258105, 282062, 329060, 306097, 274544, 260415,
           267450, 311314, 324311, 296825, 271339, 284608, 228062, 162785, 111263, 58987, 26016,
           51578, 215512, 279462, 257256, 282348, 329111, 306514, 274397, 259847, 267045,
           311791, 323739, 297453, 271344, 285047, 227655, 162922, 110554, 58886, 26243))

# step 2: create a life table using the methods from Public Health England .xlsm file
le_table <- make_life_table(data = test_data,
                            grouping_vars = c("year"),
                            age_cat_var = "ages",
                            population_var = "population_est")

# step 3: pull out the life expectancy for each group
## 3a: get le for age 0 (typical estimation used) with confidence intervals
get_le(le_table, grouping_vars = c("year"))

## 3b. get le for the age group starting at 30 without confidence intervals
get_le(le_table, start_age = 30, grouping_vars = c("year"), include_ci = FALSE)

}
