% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-management.R
\name{remove_right_empty_cols}
\alias{remove_right_empty_cols}
\title{Remove all empty columns on the right end of a data set.}
\usage{
remove_right_empty_cols(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
the inputted data frame without empty columns on the right
}
\description{
This function will drop empty columns at the far right of a data set. It moves from the right to left and will stop removing empty columns when it gets to the first column with data in it.
}
\examples{

data <- data.frame(
  col1 = sample(1:100, 100, replace = TRUE),
  col2 = NA,
  col3 = runif(100),
  col4 = sample(c("blue", "green", "yellow", "orange"), 100, replace = TRUE),
  col5 = NA,
  col6 = NA)

remove_right_empty_cols(data)
 # the function drops `col5` and `col6` but keeps `col3`:
head(data)


}
