% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc-geographies.R
\name{smc_zip_region_sort}
\alias{smc_zip_region_sort}
\title{Classify zip codes by region in San Mateo County}
\usage{
smc_zip_region_sort(data,
                           zip_col = "zip",
                           region_col = "zip_region")
}
\arguments{
\item{data}{This is the name of the dataframe with the zip code variable in it that you'd like to sort}

\item{zip_col}{(optional): This is a string that specifies the name of the zip code variable you want to sort. By default the function assumes the zipcodes are stored in a variable called \code{zip}}

\item{region_col}{(optional): This is a string to specify the name of the variable with the zip regions. By default the sorted zipcodes will be saved in a variable called \code{region_col}}
}
\value{
a dataset with a new variable for the sorted zip codes
}
\description{
This function sorts 5-digit San Mateo zip codes into county regions: north, south, mid, coastside and "not a residential zip". Non-San Mateo zipcodes will return an \code{NA} in the \code{city_clean} column.
}
\examples{
data <- data.frame(
  cty_zip = c("94015", "94403", "94303", "94019", "94128", "94110")
  ) \%>\%

data \%>\%
  smc_zip_region_sort(zip_col = "cty_zip",
                      region_col = "cty_zip_region")


}
