% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-smc.R
\name{theme_ft_smc}
\alias{theme_ft_smc}
\title{Using \code{theme_ft_smc}}
\usage{
theme_ft_smc(ft)
}
\arguments{
\item{ft}{This should be a \code{flextable} object}
}
\value{
a \code{flextable} object with custom formatting
}
\description{
This function will format a flextable object according to the San Mateo County Office of Epidemiology and Evaluation's style guide.
}
\examples{
iris \%>\%
   dplyr::slice(1:5) \%>\%
   dplyr::select(Species, everything()) \%>\%
   flextable() \%>\%
   footnote(i = 1, j = 1, value = as_paragraph("footer test")) \%>\%
   theme_ft_smc()

}
